Initial commit
This includes basic setup, and a PEG parser for the logic syntax. I've
included some basic tests for the parser, though I've made some
assumptions that might not hold (e.g. a > 2 is valid but 2 < a is not).
I'll be interested to see what throwing this at a real data set looks
like.

It would have been nice to better encode the bool/num distinction in
these types, but the Pest API requires a single output type. So even
though this should never result in a num where a bool is expected or
vice versa, we'll still need to actually encode that in the type system
during the "canonicalization" phase