Add a higher level logic representation
I think this is going to be the level of abstraction we actually
evaluate everything at. I had imagined going even one step further,
where we remove "Or" entirely and expand everything to a list of "all
these conditions are true", e.g. `a + (b | c)` becomes `["a + b", "a +
c"]`.

However, comparison ops throw a wrench in that. I think the only real
reason we would want to do that is because it could enable some highly
optimized set operations, but with comparison ops that's no longer
really an option. If performance of lookups does become an issue maybe
I'll investigate that further and just separate comparisons out on their
own.

All that said, if we're not planning on doing any optimziations around
comparing two sets, idents/terms just sorta stop existing. Fundamentally
`LEFTCLAW` is the same as `LEFTCLAW>0`, so everything becomes a
comparison.

The tests here are much more sparse. There's not a huge amount of logic
to test, and I don't want to re-test the string parsing logic here. The
only condition I'm at all concerned about is a few places where
`RANDOMELEVATORS=0` is in the code, which would underflow with the naïve
logic