Begin parsing the spoiler log.
This includes the boilerplate required for me to try to parse the logic
section of the spoiler log, as well as the grammar changes required to
get that data to parse. The following changes became clear:

- Idents can contain "-"
- There are some special idents like $StartLocation
- Some of the door names have spaces (probably only when used with
  $StartLocation)
- Comparisons can happen with special ints like $NotchCost.

I really don't know how I want to handle $NotchCost yet, but I'm going
to ignore it since logic that requires equipping two separate charms
(and thus the only time $NotchCost would come into play) is going to be
very rare. I suspect what I'll want to do is pass the IDs of the notches
in when converting from RawLogic to Condition, so I can just turn it
into an int there.