Working prototype
Ok, this gets us to the bare minimum working prototype. There are
definitely edge cases we aren't handling yet. The biggest one is that we
know nothing about costs. We also still treat notch requirements as
always met.

But in theory this should work any time that there's a single item
pickup that would work. This takes the path to a directory continaining
the RawSpoiler.json (so we can get the actual logic for a given save,
and the item locations), and TrackerDataWithoutSequenceBreaksPM.txt (so
we can get information about what they already have access to), and
looks for any single item pickup that changes the set of reachable
locations.

Since the tracker data isn't actually valid JSON (trailing comma), I had
to switch to something more permissive than serde_json. I didn't grab
json5 since it requires loading the whole file into memory, but maybe
that's not such a huge deal.