Respect item costs
This adds the logic required to check if the player can actually afford
an item, and change the logic based on that. Determining if we have
anything to output, and what information it should include got a bit
more complicated, since now pickups might just unlock an item at a
location we already had access to.

There's also one more wrinkle here, which is that the value the player
has is based on a tolerance setting in the rando. For example, if an
item requires 1 grub, the logic won't actually require you to pick up
that item unless you have 3 grubs by default (tolerance of 2).

That makes it less clear what I want to do in this case, since telling
the user to go get an item from grubfather for 1 grub will put them out
of logic. However, telling them to go get 3 grubs and then talk to
grubfather when the actual cost they can see is 1 feels equally
confusing.

For now I felt it was better to base things off the actual cost, without
tolerance. Down the line I'll probably have the output note something
like "(out of logic unless you pick up 2 additional grubs)" or similar,
and maybe even filter out any results where there aren't actually enough
reachable items to acquire it in logic.