Revert showing pickups that unlock items in shops
The answers we were giving as a result of this could be extremely wrong.
We have no concept of which items the user has already picked up, and
thus no way of knowing which grub/essence/etc to point them at. We'll
need to do something more intelligent in to figure out whether it's
possible to unlock new items in shops. And when we do that, we won't be
able to point at a specific pickup, we'll have to say "all reachable
grubs" and such