Make the thing do the thing in the browser
Ok this took comically little code. `wasm` contains a very small shim to
connect the main run function to the browser. And then www contains the
actual browser code to:

- Read the two files
- Have some checkboxes for the flags
- Turn the files into some arrays
- Call the thing

Most of the boilerplate came from the getting started guide for
wasm-pack. I tried to remove all the redundant bits but something may
have slipped in