Use `TrackerLog.txt` instead of `TrackerDataWithoutSequenceBreaksPM.txt`
I was really surprised that this was able to work without knowing which
items the user had picked up, and finally we found the reason it can't
actually be that simple. Some items, such as the cloak, would unlock
more than one location if you could pick it up multiple times.

So this puts us in the not so great situation of needing the user to
give us that information. The only place where it's easily accessible in
a machine readable format is in the actual save file. But that's in a
different directory from the spoiler log, which for the web UI would be
painful to work with. I might consider it if we could take *only* that
file and not need the spoiler log, but the save doesn't include
information like the logic for each check, and the item placements also
would be difficult to reconstruct.

So that leaves us with the files in the randomizer's log files. Nothing
here has the information I need in a machine readable format, but
TrackerLog.txt has it in a form that at least is relatively easy to
crunch through without actually parsing it, so I've gone with that.

Since we know the actual locations that have been cleared, this means
we no longer need the TrackerData file that we were using before. That
file was nice to use since it had every reachable transition already
marked. Now that we're not using that file, I have to re-implement that
logic.

Additionally, the spoiler log doesn't include the actual transition data
when transition rando isn't turned on. So we need to include the vanilla
transitions in the binary. We will need additional logic to handle
transition rando, which I have not included in this commit.

At this point our logic really is starting to look like a shitty
implementation of prolog or datalog. It'll probably be worthwhile to
investigate using a library with an actual implementation of that type
of logic.