Add support for transition randomizer
This is quite hard to test thoroughly in a reasonable amount of time,
but this appears to be all that's needed from some cursory testing. The
value will be `null` when transition rando is turned off, or an array
with that information when it's turned on.

Not all transitions get randomized all the time, and I'm not sure how
the transitions that remain vanilla get handled, so I've done what I
*think* is the most resillient, not even considering the setting and
just instead looking for a transition from spoiler log first, and then
falling back to vanilla.