Try really really hard to find the item the player picked up
Just applying the item that appears in the log is not enough for us. The
log does not actually always show the full name of the item, but rather
the name of whatever item was shown to the player. This is a problem for
us if the item is a progressive pickup, since the player could pick up
Unbreakable_Greed but the log will show Fragile_Greed. Even worse,
there's no information in the spoiler log that we could use to link
those two names together.

I don't want to hard code this information, since so far we've been able
to avoid it and that should mean this "just works" even with mods that
add items/checks we don't know about.

If we can't hard code the info, the only approach I could think of was
to look for an item with that name, and then find the item at the logged
location that has the same effects. It's ugly, but it should get the job
done.

Except even that's not quite enough. There's two edge cases I was able
to quickly find. The first is that when you have duplicated items, any
pickups after the max will be replaced with 100 Geo. So we need to
filter those out. The other is that the second white fragment piece you
pick up will be logged as "Kingsoul" not a specific fragment. Right now
I'm just assuming we always get queen's fragment first, so Kingsoul is
Kings_Fragment. This is probably not right and I'll need to figure out
how it works as I can throw more saves at it.

This is still not going to be enough to get shops working correctly.
Since we're not actually storing the item/location pair, any progressive
item where checks are enabled by pickups after the first are still going
to be shown even after they've already been picked up. I vaguely know
how I want to handle this, but I don't have a save I can use to test
yet.