Locations with items that have the same name are reachable
There are a ton of cases where a pickup is named after the item that is
there normally. Before this commit, we would stop considering that
location reachable as soon as you picked up the item with that name.

I think I originally added this !already_unlocked check becuase in my
head we would get into an infinite loop otherwise when evaluating
conditions, since most locations have themselves as conditions. That of
course doesn't happen because we didn't implement the logic that way.

We *do* end up in an infinite loop if we don't check for that when we're
unlocking all reachable locations, so we can move the check to there.