Actually track which items are picked up
This prevents items which we've picked up in the past from ever showing
up in the list again. Prior to 7f6bd90, this was mostly enforced by the
fact that we were sticking the actual location names in the "acquired"
set, and for locations with only a single item that would ensure we
wouldn't go after that item again because we wouldn't consider the
location reachable.

But stores mess that up. Stores have more than one item and are never
"cleared". So we need some way to uniquely track these items. A
(location, name) pair isn't enough, since in the case of things like
grubs the same item can appear at a location more than once. We could
store a count but now that gets incredibly hairy. We can use the index
in which the placement appears as a unique ID.

This is still going to leave a few edge cases unhandled. In particular,
we could end up marking an unaffordable item as purchased instead of a
cheap one, making us think more grubs, etc are reachable than actually
are. I'm not even sure it's actually going to be worth bothering to fix
this, since I've been promised that these IDs will be included in the
file we're reading in the next release.