Persist file handles accross visits, display whether we can autoreload
The FileSystemHandle API is only available in Chrome, so the behavior
will different in those browsers vs Firefox. We should display that
information to the user, and if we have access to a handle, stick it in
IndexedDB so they don't have to open the filepicker again next time they
visit. (This will still result in a prompt to grant access, but that's
at least easier to click).

It'd be nice if we could show the full path, but that doesn't appear to
be possible.