Handle vanilla item placements
Once I started testing a save that had the default randomizer settings,
I noticed that I was never being pointed to go to grubfather. That was
because as far as the code was concerned, there were no grubs on the
map.

In the spoiler log, there is an array called "Vanilla", which includes
anything that *wasn't* randomized. This also included transitions,
meaning we could delete our hard coded vanilla transitions!

So we now ingest this array, place the items/transitions as appropriate,
and all is well and good. Except then I immediately spotted another bug,
which is that all the tests started saying "Go pick up RIGHTSLASH from
Start"...

Which made me realize that we didn't know the knight could attack unless
those were randomized, which makes sense. Now we pick up any items that
are listed as being placed at Start, even if they don't appear in the
log.

So this means there's a few things that have changed:

- We will now start showing shops again even if the currency for that
  shop isn't randomized
- There will no longer be erroneous output that resulted from the tool
  thinking you couldn't attack sideways
- We will now show when you can access new checks that aren't randomized
  as a result of picking up an item.

That last bit isn't necessarily actually desirable, since it means we'll
tell you to pick things up even if they only unlock a geo rock or lore
tablet. But I don't see any way to avoid this since we don't have any
notion of what items are/aren't necessary, so we can't tell a grub from
a geo rock.