Use the "logic" name for a location instead of the true name for lookups
Items from Salubra are logged as "Salubra" even if they required charms
to puchase. This would lead to us failling to find the item, since
"Salubra_(Requires_Charms)" does not appear in logic, and when we parse
the picked up items we'd fail to find the charm we looked for at
Salubra.

I think we might have the same issue with Sly even after this change,
since Sly_(Key) is the logic name, and I have a hunch the tracker logs
it as just "Sly"