Properly track randomized costs including random pickups
We used to zero-out costs when determining what could be picked up since
we had no way of knowing which specific grub, essence, etc was picked
up. Now that we *are* tracking that, we should no longer zero out the
costs, since this will cause us to think we have *less* available grubs,
essence, etc as the player picks more up.

Additionally, this causes us to begin respecting the tolerance setting
again as a side effect.