Don't assume that transitions are bi-directional
Because when we unlock a transition, we "acuqire" both the transition
itself, and the thing it points to, I had previously assumed that if we
had "acquired" a transition we don't need to do so again. However, in
chaos room rando this is no longer true. Having marked a transition as
acquired does not mean that we've marked the transition it points to as
acquired. At best this could result in incorrect output, but at worst we
can end up with a situation like the test data this commit adds, where
the transition accessible from the start would normally be a dead end,
meaning we wouldn't think you can access *any* other rooms.