Switch to Parcel 2.0
There's very little support for things in Parcel 1.0 at this point.
Parcel 2.0 doesn't support wasm out of the box, but wasm-bindgen can
generate code that doesn't require the bundler to support wasm.

I also want to experiment with some newer wasm-bindgen settings that
wasm-pack doesn't support. We're not publishing to npm so wasm-pack is
doing very little for us. We can delete it and use the wasm-bindgen CLI
directly.