Switch to a faster hash algorithm and allocator
We were spending the majority of our time in the hash function, and the
second most amount of time in the allocator on web. We can switch to
FxHash for the hash function, and remove weealloc from the web side. The
wasm code is already quite big, and the allocator was not contributing
much code size.